shift list view add
